+++ extensions/lib/websocket.h
@@ -4,14 +4,19 @@
 #include <Python.h>
 
 
-inline const char *websocket_mask(const char* chunk, const char* key,
+inline PyObject* websocket_mask(const char* chunk, const char* key,
         size_t chunk_length, size_t mask_length) {
-    char *buf = (char*)chunk;
     size_t i;
+    char* buf;
+    PyObject* result = PyBytes_FromStringAndSize(NULL, chunk_length);
+    if (!result) {
+        return NULL;
+    }
+    buf = PyBytes_AS_STRING(result);
     for (i = 0; i < chunk_length; i++) {
         buf[i] = chunk[i] ^ key[i % mask_length];
     }
-    return buf;
+    return result;
 }