+++ pulsar/async/mailbox.py
@@ -237,7 +237,7 @@ class MailboxProtocol(Protocol):
 
     def _write(self, req):
         obj = pickle.dumps(req.data, protocol=2)
-        data = self._parser.encode(obj, opcode=0x2).msg
+        data = self._parser.encode(obj, opcode=2)
         try:
             self.transport.write(data)
         except socket.error: